Add comments, renamed classes to capital letter and renamed files according to classes
