+++ src/espressif/UpdateESPClass.cpp
@@ -1,6 +1,24 @@
+/*                          =======================
+============================   C/C++ SOURCE FILE   ============================
+                            =======================                       *//**
+  UpdateESPClass.cpp
+
+  Created by Onno Dirkzwager on 10.02.2019.
+  Copyright (c) 2019 IOTAppStory. All rights reserved.
+
+*///===========================================================================
+
+/*---------------------------------------------------------------------------*/
+/*                                INCLUDES                                   */
+/*---------------------------------------------------------------------------*/
+
 #include <Arduino.h>
-#include "updateESP.h"
+#include "UpdateESPClass.h"
 #include "../serialFeedback_EN.h"
+
+/*---------------------------------------------------------------------------*/
+/*                        DEFINITIONS AND MACROS                             */
+/*---------------------------------------------------------------------------*/
 
 #ifdef DEBUG_SERIAL_ENABLE
     #define dbSerialPrint(a)    Serial.print(a)
@@ -14,6 +32,28 @@
     #define dbSerialBegin(a)    do{}while(0)
 #endif
 
+/*---------------------------------------------------------------------------*/
+/*                        TYPEDEFS AND STRUCTURES                            */
+/*---------------------------------------------------------------------------*/
+
+/*---------------------------------------------------------------------------*/
+/*                                PROTOTYPES                                 */
+/*---------------------------------------------------------------------------*/
+
+/*---------------------------------------------------------------------------*/
+/*                            LOCAL VARIABLES                                */
+/*---------------------------------------------------------------------------*/
+
+UpdateESPClass UpdateESP;
+
+/*---------------------------------------------------------------------------*/
+/*                        FUNCTION IMPLEMENTATION                            */
+/*---------------------------------------------------------------------------*/
+
+/*-----------------------------------------------------------------------------
+                        UpdateESPClass prepareUpdate
+
+*///---------------------------------------------------------------------------
 bool UpdateESPClass::prepareUpdate(uint32_t upd_size, String &upd_md5, uint16_t command) {
     _upd_size   = upd_size;
     _upd_md5    = &upd_md5;
@@ -100,6 +140,10 @@ bool UpdateESPClass::prepareUpdate(uint32_t upd_size, String &upd_md5, uint16_t 
     return true;
 }
 
+/*-----------------------------------------------------------------------------
+                        UpdateESPClass update
+
+*///---------------------------------------------------------------------------
 bool UpdateESPClass::update(uint8_t *file_buf, size_t buf_size) {
     // Write the buffered bytes to the esp. If this fails, return false.
     if(!Update.write(file_buf, buf_size)){
@@ -109,8 +153,10 @@ bool UpdateESPClass::update(uint8_t *file_buf, size_t buf_size) {
     return true;
 }
 
+/*-----------------------------------------------------------------------------
+                        UpdateESPClass end
 
-
+*///---------------------------------------------------------------------------
 bool UpdateESPClass::end() {
     _upd_size = 0;
 
@@ -136,10 +182,14 @@ bool UpdateESPClass::end() {
     return true;
 }
 
+/*-----------------------------------------------------------------------------
+                        UpdateESPClass sm
 
-
+*///---------------------------------------------------------------------------
 void UpdateESPClass::sm(String *statusMessage) {
     _statusMessage = statusMessage;
 }
 
-UpdateESPClass UpdateESP;
+/*---------------------------------------------------------------------------*/
+/*                                    EOF                                    */
+/*---------------------------------------------------------------------------*/
