+++ src/espressif/UpdateNextionClass.h
@@ -1,28 +1,27 @@
 /*                          =======================
 ============================   C/C++ HEADER FILE   ============================
                             =======================                       *//**
-  UpdateClassVirt.h
+  UpdateNextionClass.h
 
   Created by Onno Dirkzwager on 10.02.2019.
   Copyright (c) 2019 IOTAppStory. All rights reserved.
 
 *///===========================================================================
 
-#ifndef __UpdateClassVirt_h__
-#define __UpdateClassVirt_h__
+#ifndef __UpdateNextionClass_h__
+#define __UpdateNextionClass_h__
 
 /*---------------------------------------------------------------------------*/
 /*                                    INCLUDES                               */
 /*---------------------------------------------------------------------------*/
 
+#include "../config.h"
+#include "UpdateClassVirt.h"
+#include <ESPNexUpload.h>
+
 /*---------------------------------------------------------------------------*/
 /*                            DEFINITIONS AND MACROS                         */
 /*---------------------------------------------------------------------------*/
-
-#define U_FLASH   0
-#define U_SPIFFS  100
-#define U_AUTH    200
-#define U_NEXTION 300
 
 /*---------------------------------------------------------------------------*/
 /*                        TYPEDEFS, CLASSES AND STRUCTURES                   */
@@ -31,11 +30,13 @@
 /*                          =======================
 ============================   CLASS DEFINITION    ============================
                             =======================                       *//**
-  UpdateClassVirt.
+  UpdateNextionClass.
 
 *//*=========================================================================*/
-class UpdateClassVirt {
+class UpdateNextionClass : public UpdateClassVirt {
 public:
+    ~UpdateNextionClass() {};
+
     virtual bool prepareUpdate(uint32_t upd_size, String &upd_md5, uint16_t command);
 
     virtual bool update(uint8_t *file_buf, size_t buf_size);
@@ -43,13 +44,20 @@ public:
     virtual bool end(void);
 
     virtual void sm(String *statusMessage);
+
+    ESPNexUpload nextion{NEXT_BAUD};
+
+private:
+    String* _statusMessage;
 };
 
 /*---------------------------------------------------------------------------*/
 /*                                GLOBAL VARIABLES                           */
 /*---------------------------------------------------------------------------*/
 
+extern UpdateNextionClass UpdateNextion;
+
 /*---------------------------------------------------------------------------*/
 /*                                    EOF                                    */
 /*---------------------------------------------------------------------------*/
-#endif // __UpdateClassVirt_h__
+#endif // __UpdateNextionClass_h__
