+++ src/espressif/configServer.h
@@ -1,42 +1,79 @@
+/*                          =======================
+============================   C/C++ HEADER FILE   ============================
+                            =======================                       *//**
+  ConfigServer.h
+
+  Created by Onno Dirkzwager on 22.11.2018.
+  Copyright (c) 2018 IOTAppStory. All rights reserved.
+
+*///===========================================================================
+
 #if defined ESP8266 || defined ESP32
-    #ifndef configServer_h
-        #define configServer_h
 
-        #include <IOTAppStory.h>
-        #ifdef ESP32
-            #include <AsyncTCP.h>                       // https://github.com/me-no-dev/AsyncTCP
-            #include <FS.h>                             // esp32 core SPIFFS library
-            #include <SPIFFS.h>
-        #elif defined  ESP8266
-            #include <ESPAsyncTCP.h>                    // https://github.com/me-no-dev/ESPAsyncTCP
-            #include <FS.h>                             // esp8266 core SPIFFS library
-            #define FILE_WRITE  "w"
-            #define FILE_APPEND "a"
+#ifndef __ConfigServer_h__
+#define __ConfigServer_h__
 
-        #endif
-        #include <ESPAsyncWebServer.h>                  // https://github.com/me-no-dev/ESPAsyncWebServer
+/*---------------------------------------------------------------------------*/
+/*                                    INCLUDES                               */
+/*---------------------------------------------------------------------------*/
 
-        struct configStruct;
-        class IOTAppStory;
+#include "IOTAppStory.h"
 
+#ifdef ESP32
+    #include <AsyncTCP.h> // https://github.com/me-no-dev/AsyncTCP
+    #include <FS.h> // esp32 core SPIFFS library
+    #include <SPIFFS.h>
+#elif defined  ESP8266 // ESP32
+    #include <ESPAsyncTCP.h> // https://github.com/me-no-dev/ESPAsyncTCP
+    #include <FS.h> // esp8266 core SPIFFS library
+    #define FILE_WRITE  "w"
+    #define FILE_APPEND "a"
+#endif // ESP8266
 
-        class configServer {
-            public:
-                configServer(IOTAppStory &ias, configStruct &config);
-                void run();
+#include <ESPAsyncWebServer.h> // https://github.com/me-no-dev/ESPAsyncWebServer
 
-            private:
-                IOTAppStory* _ias;
-                configStruct* _config;
-                std::unique_ptr<AsyncWebServer>     server;
+/*---------------------------------------------------------------------------*/
+/*                            DEFINITIONS AND MACROS                         */
+/*---------------------------------------------------------------------------*/
 
-                bool _tryToConn         = false;        // is the wifi connector busy? (trying to connect)
-                bool _connFail          = false;        // did the last connection attempt faile
-                bool _connChangeMode    = false;        // flag to notify the loop to change from AP to STA mode
+/*---------------------------------------------------------------------------*/
+/*                        TYPEDEFS, CLASSES AND STRUCTURES                   */
+/*---------------------------------------------------------------------------*/
 
-                void onUpload(AsyncWebServerRequest *request, String filename, size_t index, uint8_t *data, size_t len, bool final);
-                void hdlReturn(AsyncWebServerRequest *request, String retHtml, String type = "text/html");
+struct configStruct;
+class IOTAppStory;
 
-        };
-    #endif
-#endif
+/*                          =======================
+============================   CLASS DEFINITION    ============================
+                            =======================                       *//**
+  ConfigServer.
+
+*//*=========================================================================*/
+class ConfigServer {
+public:
+    ConfigServer(IOTAppStory& ias, configStruct& config);
+    void run();
+
+private:
+    IOTAppStory* _ias;
+    configStruct* _config;
+    std::unique_ptr<AsyncWebServer>     server;
+
+    bool _tryToConn         = false;        // is the wifi connector busy? (trying to connect)
+    bool _connFail          = false;        // did the last connection attempt faile
+    bool _connChangeMode    = false;        // flag to notify the loop to change from AP to STA mode
+
+    void onUpload(AsyncWebServerRequest *request, String filename, size_t index, uint8_t *data, size_t len, bool final);
+    void hdlReturn(AsyncWebServerRequest *request, String retHtml, String type = "text/html");
+
+};
+
+/*---------------------------------------------------------------------------*/
+/*                                GLOBAL VARIABLES                           */
+/*---------------------------------------------------------------------------*/
+
+/*---------------------------------------------------------------------------*/
+/*                                    EOF                                    */
+/*---------------------------------------------------------------------------*/
+#endif // __ConfigServer_h__
+#endif // ESP8266 || defined ESP32
