+++ src/espressif/esp32/boardInfo.h
@@ -1,18 +1,54 @@
-#ifdef  ESP32
-    #ifndef boardInfo_h
-        #define boardInfo_h
+/*                          =======================
+============================   C/C++ HEADER FILE   ============================
+                            =======================                       *//**
+  BoardInfo.h
 
-        #include <Preferences.h>
+  Created by Onno Dirkzwager on 22.11.2018.
+  Copyright (c) 2018 IOTAppStory. All rights reserved.
 
-        class boardInfo {
-            public:
-                boardInfo(int &bootTimes, char &boardMode);
-                void read();
-                void write();
+*///===========================================================================
 
-            private:
-                int* _bootTimes;
-                char* _boardMode;
-        };
-    #endif
-#endif
+#ifdef ESP32
+#ifndef __BoardInfo_h__
+#define __BoardInfo_h__
+
+/*---------------------------------------------------------------------------*/
+/*                                    INCLUDES                               */
+/*---------------------------------------------------------------------------*/
+
+#include <Preferences.h>
+
+/*---------------------------------------------------------------------------*/
+/*                            DEFINITIONS AND MACROS                         */
+/*---------------------------------------------------------------------------*/
+
+/*---------------------------------------------------------------------------*/
+/*                        TYPEDEFS, CLASSES AND STRUCTURES                   */
+/*---------------------------------------------------------------------------*/
+
+/*                          =======================
+============================   CLASS DEFINITION    ============================
+                            =======================                       *//**
+  BoardInfo.
+
+*//*=========================================================================*/
+class BoardInfo {
+public:
+    BoardInfo(int &bootTimes, char &boardMode);
+    void read();
+    void write();
+
+private:
+    int* _bootTimes;
+    char* _boardMode;
+};
+
+/*---------------------------------------------------------------------------*/
+/*                                GLOBAL VARIABLES                           */
+/*---------------------------------------------------------------------------*/
+
+/*---------------------------------------------------------------------------*/
+/*                                    EOF                                    */
+/*---------------------------------------------------------------------------*/
+#endif // __BoardInfo_h__
+#endif // ESP32
