+++ src/espressif/esp8266/CallServer.h
@@ -1,0 +1,124 @@
+/*                          =======================
+============================   C/C++ HEADER FILE   ============================
+                            =======================                       *//**
+  CallServer.h
+
+  Created by Onno Dirkzwager on 10.02.2019.
+  Copyright (c) 2019 IOTAppStory. All rights reserved.
+
+*///===========================================================================
+
+#ifdef ESP8266
+#ifndef __CallServer_h__
+#define __CallServer_h__
+
+/*---------------------------------------------------------------------------*/
+/*                                    INCLUDES                               */
+/*---------------------------------------------------------------------------*/
+
+#include "IOTAppStory.h"
+#include <ESP8266WiFi.h>
+
+#if HTTPS == true
+    #include <WiFiClientSecureBearSSL.h>
+    #include <time.h>
+#endif // HTTPS
+
+/*---------------------------------------------------------------------------*/
+/*                            DEFINITIONS AND MACROS                         */
+/*---------------------------------------------------------------------------*/
+
+#if OTA_UPD_CHECK_NEXTION == true
+    #define U_NEXTION 300
+#endif
+
+#define U_LOGGER 400
+
+/*---------------------------------------------------------------------------*/
+/*                        TYPEDEFS, CLASSES AND STRUCTURES                   */
+/*---------------------------------------------------------------------------*/
+
+struct configStruct;
+struct firmwareStruct;
+
+/*                          =======================
+============================   CLASS DEFINITION    ============================
+                            =======================                       *//**
+  CallServer.
+
+*//*=========================================================================*/
+class CallServer {
+public:
+    CallServer(configStruct &config, int command = U_FLASH);
+
+    Stream& getStream(firmwareStruct *firmwareStruct);
+
+    bool get(const char* url, String args);
+
+    void sm(String *statusMessage);
+
+private:
+    #if HTTPS == true
+        #if HTTPS_8266_TYPE == CERTIFICATE
+            BearSSL::WiFiClientSecure _client;
+        #elif HTTPS_8266_TYPE == FNGPRINT
+            WiFiClientSecure _client;
+        #endif
+    #else
+        WiFiClient _client;
+    #endif
+    String *_statusMessage;
+
+    configStruct* _config;
+    int _command;
+    const char* _callHost = OTA_HOST;       // ota update host | Set in config.h
+    const char* _callFile = OTA_UPD_FILE;   // file at host that handles esp updates
+};
+
+/*---------------------------------------------------------------------------*/
+/*                                GLOBAL VARIABLES                           */
+/*---------------------------------------------------------------------------*/
+
+#if HTTPS_CERT_STORAGE == ST_PROGMEM
+const uint8_t ROOT_CA[] PROGMEM = \
+    "-----BEGIN CERTIFICATE-----\n" \
+    "MIIF2DCCA8CgAwIBAgIQTKr5yttjb+Af907YWwOGnTANBgkqhkiG9w0BAQwFADCB\n" \
+    "hTELMAkGA1UEBhMCR0IxGzAZBgNVBAgTEkdyZWF0ZXIgTWFuY2hlc3RlcjEQMA4G\n" \
+    "A1UEBxMHU2FsZm9yZDEaMBgGA1UEChMRQ09NT0RPIENBIExpbWl0ZWQxKzApBgNV\n" \
+    "BAMTIkNPTU9ETyBSU0EgQ2VydGlmaWNhdGlvbiBBdXRob3JpdHkwHhcNMTAwMTE5\n" \
+    "MDAwMDAwWhcNMzgwMTE4MjM1OTU5WjCBhTELMAkGA1UEBhMCR0IxGzAZBgNVBAgT\n" \
+    "EkdyZWF0ZXIgTWFuY2hlc3RlcjEQMA4GA1UEBxMHU2FsZm9yZDEaMBgGA1UEChMR\n" \
+    "Q09NT0RPIENBIExpbWl0ZWQxKzApBgNVBAMTIkNPTU9ETyBSU0EgQ2VydGlmaWNh\n" \
+    "dGlvbiBBdXRob3JpdHkwggIiMA0GCSqGSIb3DQEBAQUAA4ICDwAwggIKAoICAQCR\n" \
+    "6FSS0gpWsawNJN3Fz0RndJkrN6N9I3AAcbxT38T6KhKPS38QVr2fcHK3YX/JSw8X\n" \
+    "pz3jsARh7v8Rl8f0hj4K+j5c+ZPmNHrZFGvnnLOFoIJ6dq9xkNfs/Q36nGz637CC\n" \
+    "9BR++b7Epi9Pf5l/tfxnQ3K9DADWietrLNPtj5gcFKt+5eNu/Nio5JIk2kNrYrhV\n" \
+    "/erBvGy2i/MOjZrkm2xpmfh4SDBF1a3hDTxFYPwyllEnvGfDyi62a+pGx8cgoLEf\n" \
+    "Zd5ICLqkTqnyg0Y3hOvozIFIQ2dOciqbXL1MGyiKXCJ7tKuY2e7gUYPDCUZObT6Z\n" \
+    "+pUX2nwzV0E8jVHtC7ZcryxjGt9XyD+86V3Em69FmeKjWiS0uqlWPc9vqv9JWL7w\n" \
+    "qP/0uK3pN/u6uPQLOvnoQ0IeidiEyxPx2bvhiWC4jChWrBQdnArncevPDt09qZah\n" \
+    "SL0896+1DSJMwBGB7FY79tOi4lu3sgQiUpWAk2nojkxl8ZEDLXB0AuqLZxUpaVIC\n" \
+    "u9ffUGpVRr+goyhhf3DQw6KqLCGqR84onAZFdr+CGCe01a60y1Dma/RMhnEw6abf\n" \
+    "Fobg2P9A3fvQQoh/ozM6LlweQRGBY84YcWsr7KaKtzFcOmpH4MN5WdYgGq/yapiq\n" \
+    "crxXStJLnbsQ/LBMQeXtHT1eKJ2czL+zUdqnR+WEUwIDAQABo0IwQDAdBgNVHQ4E\n" \
+    "FgQUu69+Aj36pvE8hI6t7jiY7NkyMtQwDgYDVR0PAQH/BAQDAgEGMA8GA1UdEwEB\n" \
+    "/wQFMAMBAf8wDQYJKoZIhvcNAQEMBQADggIBAArx1UaEt65Ru2yyTUEUAJNMnMvl\n" \
+    "wFTPoCWOAvn9sKIN9SCYPBMtrFaisNZ+EZLpLrqeLppysb0ZRGxhNaKatBYSaVqM\n" \
+    "4dc+pBroLwP0rmEdEBsqpIt6xf4FpuHA1sj+nq6PK7o9mfjYcwlYRm6mnPTXJ9OV\n" \
+    "2jeDchzTc+CiR5kDOF3VSXkAKRzH7JsgHAckaVd4sjn8OoSgtZx8jb8uk2Intzna\n" \
+    "FxiuvTwJaP+EmzzV1gsD41eeFPfR60/IvYcjt7ZJQ3mFXLrrkguhxuhoqEwWsRqZ\n" \
+    "CuhTLJK7oQkYdQxlqHvLI7cawiiFwxv/0Cti76R7CZGYZ4wUAc1oBmpjIXUDgIiK\n" \
+    "boHGhfKppC3n9KUkEEeDys30jXlYsQab5xoq2Z0B15R97QNKyvDb6KkBPvVWmcke\n" \
+    "jkk9u+UJueBPSZI9FoJAzMxZxuY67RIuaTxslbH9qh17f4a+Hg4yRvv7E491f0yL\n" \
+    "S0Zj/gA0QHDBw7mh3aZw4gSzQbzpgJHqZJx64SIDqZxubw5lT2yHh17zbqD5daWb\n" \
+    "QOhTsiedSrnAdyGN/4fy3ryM7xfft0kL0fJuMAsaDk527RH89elWsn2/x20Kk4yl\n" \
+    "0MC2Hb46TpSi125sC8KKfPog88Tk5c0NqMuRkrF8hey1FGlmDoLnzc7ILaZRfyHB\n" \
+    "NVOFBkpdn627G190\n" \
+    "-----END CERTIFICATE-----\n";
+#endif
+
+/*---------------------------------------------------------------------------*/
+/*                                    EOF                                    */
+/*---------------------------------------------------------------------------*/
+#endif // __CallServer_h__
+#endif // ESP8266
