+++ src/includes.h
@@ -1,3 +1,17 @@
+/*                          =======================
+============================   C/C++ HEADER FILE   ============================
+                            =======================                       *//**
+  includes.h
+
+  Created by Onno Dirkzwager on 10.02.2019.
+  Copyright (c) 2019 IOTAppStory. All rights reserved.
+
+*///===========================================================================
+
+/*---------------------------------------------------------------------------*/
+/*                                    INCLUDES                               */
+/*---------------------------------------------------------------------------*/
+
 #ifdef ESP32
     #include <rom/spi_flash.h>                  // enable flash chip id from the SDK
     #include <WiFi.h>
@@ -5,22 +19,22 @@
     #include "espressif/WiFiConnector.h"
     #include <ESPmDNS.h>
 
-    #include "espressif/esp32/boardInfo.h"
-    #include "espressif/esp32/callServer_WiFiClientSecure.h"
-    #include "espressif/configServer.h"
+    #include "espressif/esp32/BoardInfo.h"
+    #include "espressif/esp32/CallServer.h"
+    #include "espressif/ConfigServer.h"
 
     #include "espressif/UpdateClassVirt.h"
     #include <Update.h>
-    #include "espressif/updateESP.h"
+    #include "espressif/UpdateESPClass.h"
     #if OTA_UPD_CHECK_NEXTION == true
-        #include "espressif/updateNextion.h"
-    #endif
+        #include "espressif/UpdateNextionClass.h"
+    #endif // OTA_UPD_CHECK_NEXTION
 
     #include <HTTPClient.h>
     #include <DNSServer.h>
     #include <EEPROM.h>
     #include <ESPAsyncWebServer.h>              // https://github.com/me-no-dev/ESPAsyncWebServer
-#endif
+#endif // ESP32
 
 #ifdef  ESP8266
     #include <ESP8266WiFi.h>
@@ -28,21 +42,21 @@
     #include "espressif/WiFiConnector.h"
     #include <ESP8266mDNS.h>
 
-    #include "espressif/esp8266/boardInfo.h"
-    #include "espressif/esp8266/callServer_WiFiClientSecure.h"
-    #include "espressif/configServer.h"
+    #include "espressif/esp8266/BoardInfo.h"
+    #include "espressif/esp8266/CallServer.h"
+    #include "espressif/ConfigServer.h"
 
     #include "espressif/UpdateClassVirt.h"
-    #include "espressif/updateESP.h"
+    #include "espressif/UpdateESPClass.h"
     #if OTA_UPD_CHECK_NEXTION == true
-        #include "espressif/updateNextion.h"
-    #endif
+        #include "espressif/UpdateNextionClass.h"
+    #endif // OTA_UPD_CHECK_NEXTION
 
     #include <ESP8266HTTPClient.h>
     #include <DNSServer.h>
     #include <EEPROM.h>
     #include <ESPAsyncWebServer.h>              // https://github.com/me-no-dev/ESPAsyncWebServer
-#endif
+#endif // ESP8266
 
 #ifdef ARDUINO_SAMD_VARIANT_COMPLIANCE
     //#WIFI_MULTI false
@@ -58,4 +72,20 @@
     //#include "ardmkr/otaUpdate.h"
 
     #include <FlashAsEEPROM.h>
-#endif
+#endif // ARDUINO_SAMD_VARIANT_COMPLIANCE
+
+/*---------------------------------------------------------------------------*/
+/*                            DEFINITIONS AND MACROS                         */
+/*---------------------------------------------------------------------------*/
+
+/*---------------------------------------------------------------------------*/
+/*                        TYPEDEFS, CLASSES AND STRUCTURES                   */
+/*---------------------------------------------------------------------------*/
+
+/*---------------------------------------------------------------------------*/
+/*                                GLOBAL VARIABLES                           */
+/*---------------------------------------------------------------------------*/
+
+/*---------------------------------------------------------------------------*/
+/*                                    EOF                                    */
+/*---------------------------------------------------------------------------*/
